Refactored code and added unit test cases for services.

+++ src/main/java/com/medianocean/assessment/repository/TeamRepository.java
@@ -14,5 +14,5 @@ public interface TeamRepository {
 	
 	Team findById(int id);
 	
-	long countTeams();
+	int countTeams();
 }
